Update to save translate state.
