+++ src/app/cmds/translate/translate.state.ts
@@ -1,0 +1,13 @@
+import { TranslatorLanguage } from "../../components/translator_language/translator_language.interface";
+
+export type TranslateState = {
+  autoSize: boolean;
+  sourceLanguage: TranslatorLanguage;
+  fontReplacementState: {
+    language: TranslatorLanguage;
+    fontName: string;
+    font: FontName;
+    isChecked: boolean;
+    isVisible: boolean;
+  }[];
+};
