+++ src/app/components/config/config.service.ts
@@ -1,15 +1,21 @@
-import { ConfigEntity } from "./config";
-import { Config, ConfigRepository, ConfigService } from "./config.interface";
+import { TranslateState } from "../../cmds/translate/translate.state";
+import { ConfigRepository, ConfigService } from "./config.interface";
 
 export class ConfigServiceImpl implements ConfigService {
   constructor(private configRepository: ConfigRepository) {}
 
-  public async get(): Promise<Config> {
-    return (
-      (await this.configRepository.get()) ||
-      new ConfigEntity({
-        googleAPIKey: "",
-      })
-    );
+  public async getTranslateState(): Promise<TranslateState | undefined> {
+    const config = await this.configRepository.get();
+    return config?.translateState;
+  }
+
+  public async setTranslateState(
+    translateState: TranslateState
+  ): Promise<void> {
+    const config = await this.configRepository.get();
+    this.configRepository.set({
+      ...config,
+      translateState,
+    });
   }
 }
