--- src/app/components/config/config.ts
@@ -1,9 +1,0 @@
-import { Config } from "./config.interface";
-
-export class ConfigEntity implements Config {
-  public googleAPIKey: string;
-
-  constructor({ googleAPIKey }: Config) {
-    this.googleAPIKey = googleAPIKey;
-  }
-}
