+++ src/ui/translate/index.html
@@ -36,7 +36,6 @@
           <input
             id="fontReplacementCheckbox"
             type="checkbox"
-            checked
             style="margin: 9px 8px 8px 4px"
           />
           <!-- Search -->
